+++ src/lib.rs
@@ -1,7 +1,8 @@
 pub use mlg::{
-    CheckResult, DebugResult, ExportResult, InitResult, LspResult, ReleaseResult, VersionResult,
-    ViewResult, WhteRbtObjResult, check, check_diagnostics_report, check_diagnostics_schema, debug,
-    export, init, lsp, release, version, view, whte_rbt_obj,
+    CheckResult, CleanResult, DebugResult, ExportResult, InitResult, LspResult, ReleaseResult,
+    VersionResult, ViewResult, WhteRbtObjResult, check, check_diagnostics_report,
+    check_diagnostics_schema, clean, debug, export, init, lsp, release, version, view,
+    whte_rbt_obj,
 };
 pub mod backend;
 pub mod cli;